renamed "springadventures" to "networkhandler"
